refactor(interactor): move internal/interactor to pkg/interactor (SRD-012 M1a)
First leaf of the ADR-012 layering relocation. internal/interactor already
imports only public pkg/model/* and has no internal-only implementor, so it
moves wholesale to the public pkg/interactor (package name unchanged).

Re-point the three importers (internal/renv, internal/instance, the model's
user_task.go) and the .mockery.yaml package key; regenerate mocks (the mock
package mockinteractor is unchanged, so test call-sites are untouched).

Behaviour-preserving; user_task.go no longer imports internal/interactor.

Implements SRD-012 v.1 FR-4 (part of M1).
